<feat>(TaskListComponent): Adds feature of being able to change status of a task and seprating done tasks

+++ src/main/webapp/app/tasks/task-list/task-list.component.ts
@@ -1,7 +1,7 @@
-import { ChangeDetectionStrategy, Component, EventEmitter, Inject, Input, Output } from '@angular/core';
-
-import { Task } from '../task';
-import { TaskService } from '../task.service';
+import { CdkDragDrop, moveItemInArray, transferArrayItem } from '@angular/cdk/drag-drop';
+import { AfterContentChecked, ChangeDetectionStrategy, Component, EventEmitter, Inject, Input, Output } from '@angular/core';
+import { LocalTaskService } from '../local-task.service';
+import { Task, TaskStatus } from '../task';
 
 /**
  * A list of tiny tasks.
@@ -12,17 +12,77 @@ import { TaskService } from '../task.service';
   styleUrls: ['./task-list.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class TaskListComponent {
+export class TaskListComponent implements AfterContentChecked {
 
   @Input() tasks: Task[];
-
   @Output() deleted: EventEmitter<Task> = new EventEmitter();
+  @Output() markedAsDone: EventEmitter<Task> = new EventEmitter();
+  @Output() deleteedAllDoneTasks: EventEmitter<Task> = new EventEmitter();
 
-  constructor(@Inject('TaskService') private taskService: TaskService) { }
+  public doneTasks: Array<Task>;
+  public toDoTasks: Array<Task>;
+  public statuses: Array<string>;
 
-  delete(task: Task): void {
-    this.taskService.delete(task.id).subscribe(() => {
-      this.deleted.emit(task);
+  constructor(@Inject('TaskService') private taskService: LocalTaskService) {}
+
+  /**
+   * Angular lifecycle hook, called after any @Input propery is changed
+   * Used to get task bifergation using task status
+   */
+  ngAfterContentChecked() {
+    if (this.tasks) {
+      this.doneTasks = this.tasks.filter( task => task.status === TaskStatus.Done );
+      this.toDoTasks = this.tasks.filter( task => task.status !== TaskStatus.Done );
+    }
+    this.statuses = Object.values(TaskStatus);
+  }
+
+  /**
+   * Callback for delete request of a task, removes task from local storage using id
+   * @param task task object which needs to be deleted
+   */
+  public delete(task: Task): void {
+    if (confirm('Are you sure you want to delete this task?')) {
+      this.taskService.delete(task.id).subscribe(() => {
+        this.deleted.emit(task);
+      });
+    }
+  }
+
+  /**
+   * Callback for change event of status dropdown, changes status of the given task
+   * @param task task for which status needs to be updated
+   * @param event change event reference to get selected value from dropdown
+   */
+  public changeStatus(task: Task, event: any) {
+    this.taskService.changeStatus(task.id, event.value).subscribe(() => {
+      this.markedAsDone.emit(task);
+    });
+  }
+
+  /**
+   * Callback for deleting request for all done task
+   * Removes all done task from local storage
+   */
+  public deleteAllDoneTasks(): void {
+    if (confirm('Are you sure you want to delete all done tasks?')) {
+      this.taskService.deleteAllTasks().subscribe(() => {
+        this.deleteedAllDoneTasks.emit();
+      });
+    }
+  }
+
+  drop(event: CdkDragDrop<string[]>) {
+    if (event.previousContainer === event.container) {
+      moveItemInArray(event.container.data, event.previousIndex, event.currentIndex);
+    } else {
+      transferArrayItem(event.previousContainer.data,
+                        event.container.data,
+                        event.previousIndex,
+                        event.currentIndex);
+    }
+    this.doneTasks.forEach((task: Task) => {
+        task.status = TaskStatus.Done;
     });
   }
 }
